fix bugs in op_only_trace_timestamp

+++ clients/drcachesim/tracer/tracer.h
@@ -113,7 +113,7 @@ typedef struct {
     syscall_pt_trace_t syscall_pt_trace;
 #endif
     uint64_t cur_win_size;
-    uint64_t cur_win_id;
+    uint32_t cur_win_id;
 } per_thread_t;
 
 /* Allocated TLS slot offsets */
